[PyOV] Initial support for type hinting using stub `.pyi` files (#29139)
### Turning on type hinting in VSC:
 - Install Python and Pylance extensions
- Turn on type checking as described in
https://docs.pydantic.dev/latest/integrations/visual_studio_code/#configure-pylance
 - The stubs should show hints for OpenVINO API:

![image](https://github.com/user-attachments/assets/f211ef90-ccf0-4bd0-a553-f31483ad7950)

### Details:
 - Work in progress, **draft is for internal Python team review only**
 - Fix or skip errors thrown during stub generation
- Fix some of the errors thrown by `pylance` when reading .pyi files
(there's quite a few of them, in my opinion the rest has to be fixed
separately in different PRs)
- Adjust CMake scripts for wheel building dependencies and for fdupes
check
 - Add script with xfails, which will be used for stub generation
- This PR is called "initial support" due to numerous leftover errors
reported by `pylance`, such as:

![image](https://github.com/user-attachments/assets/7c8d0740-cfe3-45d4-b564-ed93489a8a0d)
In my opinion the rest should be fixed systematically and in scope of
other PRs, mostly due to variety of errors, which often require case by
case analysis.

### TBD:
- ~~The workflow for updating stubs has to be decided on. In current
implementation, the stubs need to be generated manually and committed.
They are also stored in the repository. Maybe they should be generated
just before wheel building (automatically) and not uploaded to GitHub?~~
**The current plan is to automatically generate and commit `.pyi` files
using pre-commit git hook.**
- ~~If we decide to keep `.pyi` files like GenAI does, we'll need a CI
check to assert that they're up to date, similar to GenAI:
https://github.com/openvinotoolkit/openvino.genai/pull/1214~~ - **will
be continued in an epic specific to PyAPI stubs introduction**
- ~~Should the `generate_pyapi_stubs.sh` script be in
`openvino/scripts`?~~ **Moved to Python directory**
- ~~Is the current implementation sufficient for initial support or
should we first go through the modules and fix all the `pylance`
errors?~~

### To do:
 - Verify the `.bat` script is working correctly

### Tickets:
 - CVS-157373

---------

Signed-off-by: p-wysocki <przemyslaw.wysocki@intel.com>
Co-authored-by: Anastasia Kuporosova <anastasia.kuporosova@intel.com>