[TRANSFORMATIONS] Fix implicit conversion of ov::Node with multiple outputs to ov::Output (#29780)
### Details:
In the IndirectSDPAOpt transformation, 1st input to the KVCache node may
appear to be a
VariadicSplit node with multiple outputs. Using the default output of
the node is incorrect and leads to an exception if used as an input to
another node because of an implicit conversion to ov::Output. Use the
output of the VariadicSplit node explicitly to avoid the exception.

### Tickets:
- [CVS-163937](https://jira.devtools.intel.com/browse/CVS-163937)

Signed-off-by: Andrii Staikov <andrii.staikov@intel.com>